+++ src/library/DIPS.Mobile.UI/Components/Chips/iOS/ChipHandler.cs
@@ -18,6 +18,7 @@ namespace DIPS.Mobile.UI.Components.Chips;
 public partial class ChipHandler : ViewHandler<Chip, UIButton>
 {
     internal Button m_button;
+    private IDisposable m_observer;
 
     protected override UIButton CreatePlatformView()
     {
@@ -29,7 +30,6 @@ public partial class ChipHandler : ViewHandler<Chip, UIButton>
     {
         base.ConnectHandler(platformView);
         m_button.TextColor = Colors.GetColor(ColorName.color_system_black);
-
         // Here we style the button as close as possible to native compact datepicker in iOS
         // We do not use the design system here so this does not diverge at a later point
         m_button.FontSize = 17;
@@ -37,7 +37,6 @@ public partial class ChipHandler : ViewHandler<Chip, UIButton>
         m_button.Padding = new Thickness(12, 6, 12, 6);
         platformView.AddGestureRecognizer(new ChipGestureRecognizer(this));
     }
-
     private static partial void MapTitle(ChipHandler handler, Chip chip)
     {
         handler.m_button.Text = chip.Title;
@@ -82,7 +81,7 @@ public partial class ChipHandler : ViewHandler<Chip, UIButton>
     
     private static partial void MapCornerRadius(ChipHandler handler, Chip arg2)
     {
-        handler.m_button.CornerRadius = handler.VirtualView.CornerRadius;
+        handler.PlatformView.Layer.CornerRadius = handler.VirtualView.CornerRadius;
     }
 
     private static void ShiftImageToTheRight(ChipHandler handler, UIButton uiButton)